requirements.txt
Flask==3.0.0
tweepy==4.14.0
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.0 